requirements.txt

mysql-connector-python 9.1.0
opencv-contrib-python  4.10.0.84
pillow                 11.0.0